+++ compiler-core/src/call_graph/into_dependency_order_tests.rs
@@ -1,51 +1,32 @@
 use super::*;
-use crate::ast::{Arg, ExternalFunction, Function, ModuleFunction, TypeAst};
+use crate::ast::{Arg, Function};
 use smol_str::SmolStr;
 
-enum Input {
-    Internal(&'static str, &'static [&'static str], &'static str),
-    External(&'static str),
-}
+type Input = (&'static str, &'static [&'static str], &'static str);
 
 fn parse_and_order(functions: &[Input]) -> Result<Vec<Vec<SmolStr>>, Error> {
     let functions = functions
         .iter()
-        .map(|input| match input {
-            Input::Internal(name, arguments, src) => ModuleFunction::Internal(Function {
-                name: name.into(),
-                arguments: arguments
-                    .iter()
-                    .map(|name| Arg {
-                        names: crate::ast::ArgNames::Named { name: name.into() },
-                        location: Default::default(),
-                        annotation: None,
-                        type_: (),
-                    })
-                    .collect_vec(),
-                body: crate::parse::parse_statement_sequence(src).expect("syntax error"),
-                location: Default::default(),
-                return_annotation: None,
-                public: true,
-                end_position: src.len() as u32,
-                return_type: (),
-                documentation: None,
-                external_erlang: None,
-                external_javascript: None,
-            }),
-            Input::External(name) => ModuleFunction::External(ExternalFunction {
-                name: name.into(),
-                arguments: vec![],
-                module: "themodule".into(),
-                fun: name.into(),
-                location: Default::default(),
-                public: true,
-                return_: TypeAst::Hole {
+        .map(|(name, arguments, src)| Function {
+            name: name.into(),
+            arguments: arguments
+                .iter()
+                .map(|name| Arg {
+                    names: crate::ast::ArgNames::Named { name: name.into() },
                     location: Default::default(),
-                    name: "_".into(),
-                },
-                return_type: (),
-                documentation: None,
-            }),
+                    annotation: None,
+                    type_: (),
+                })
+                .collect_vec(),
+            body: crate::parse::parse_statement_sequence(src).expect("syntax error"),
+            location: Default::default(),
+            return_annotation: None,
+            public: true,
+            end_position: src.len() as u32,
+            return_type: (),
+            documentation: None,
+            external_erlang: None,
+            external_javascript: None,
         })
         .collect_vec();
     Ok(into_dependency_order(functions)?
@@ -53,7 +34,7 @@ fn parse_and_order(functions: &[Input]) -> Result<Vec<Vec<SmolStr>>, Error> {
         .map(|level| {
             level
                 .into_iter()
-                .map(|function| function.name().clone())
+                .map(|function| function.name.clone())
                 .collect_vec()
         })
         .collect())
@@ -63,7 +44,7 @@ fn parse_and_order(functions: &[Input]) -> Result<Vec<Vec<SmolStr>>, Error> {
 fn empty() {
     let functions = [];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         Vec::<Vec<SmolStr>>::new()
     );
 }
@@ -71,14 +52,14 @@ fn empty() {
 #[test]
 fn no_deps() {
     let functions = [
-        Input::External("a"),
-        Input::Internal("b", &[], r#""ok""#),
-        Input::Internal("c", &[], r#"1"#),
-        Input::Internal("d", &[], r#"1.0"#),
-        Input::Internal("e", &[], r#"todo"#),
+        ("a", [].as_slice(), "1"),
+        ("b", [].as_slice(), r#""ok""#),
+        ("c", [].as_slice(), r#"1"#),
+        ("d", [].as_slice(), r#"1.0"#),
+        ("e", [].as_slice(), r#"todo"#),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["a"], vec!["b"], vec!["c"], vec!["d"], vec!["e"]]
     );
 }
@@ -86,12 +67,12 @@ fn no_deps() {
 #[test]
 fn one_dep() {
     let functions = [
-        Input::External("a"),
-        Input::Internal("b", &[], r#"c"#),
-        Input::Internal("c", &[], r#"0"#),
+        ("a", [].as_slice(), "1"),
+        ("b", [].as_slice(), r#"c"#),
+        ("c", [].as_slice(), r#"0"#),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["a"], vec!["c"], vec!["b"]]
     );
 }
@@ -99,12 +80,12 @@ fn one_dep() {
 #[test]
 fn unknown_vars() {
     let functions = [
-        Input::External("a"),
-        Input::Internal("b", &[], r#"Nil"#),
-        Input::Internal("c", &[], r#"Ok"#),
+        ("a", [].as_slice(), "1"),
+        ("b", [].as_slice(), r#"Nil"#),
+        ("c", [].as_slice(), r#"Ok"#),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["a"], vec!["b"], vec!["c"]]
     );
 }
@@ -112,12 +93,12 @@ fn unknown_vars() {
 #[test]
 fn calling_function() {
     let functions = [
-        Input::Internal("a", &[], r#"b()"#),
-        Input::Internal("b", &[], r#"c(1, 2)"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"b()"#),
+        ("b", [].as_slice(), r#"c(1, 2)"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["c"], vec!["b"], vec!["a"]]
     );
 }
@@ -125,12 +106,12 @@ fn calling_function() {
 #[test]
 fn ref_in_call_argument() {
     let functions = [
-        Input::Internal("a", &[], r#"c(1, b())"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"c(1, b())"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["c"], vec!["a"]]
     );
 }
@@ -138,12 +119,12 @@ fn ref_in_call_argument() {
 #[test]
 fn sequence() {
     let functions = [
-        Input::Internal("a", &[], r#"c({ 1 2 b })"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"c({ 1 2 b })"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["c"], vec!["a"]]
     );
 }
@@ -151,12 +132,12 @@ fn sequence() {
 #[test]
 fn tuple() {
     let functions = [
-        Input::Internal("a", &[], r#"#(b, c, 1)"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"#(b, c, 1)"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["c"], vec!["a"]]
     );
 }
@@ -164,12 +145,12 @@ fn tuple() {
 #[test]
 fn pipeline() {
     let functions = [
-        Input::Internal("a", &[], r#"1 |> b |> c"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"1 |> b |> c"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["c"], vec!["a"]]
     );
 }
@@ -177,12 +158,12 @@ fn pipeline() {
 #[test]
 fn list() {
     let functions = [
-        Input::Internal("a", &[], r#"[b, b, c, 1]"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"[b, b, c, 1]"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["c"], vec!["a"]]
     );
 }
@@ -190,12 +171,12 @@ fn list() {
 #[test]
 fn list_spread() {
     let functions = [
-        Input::Internal("a", &[], r#"[b, b, ..c]"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"[b, b, ..c]"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["c"], vec!["a"]]
     );
 }
@@ -203,12 +184,12 @@ fn list_spread() {
 #[test]
 fn record_access() {
     let functions = [
-        Input::External("a"),
-        Input::Internal("b", &[], r#"b().wibble"#),
-        Input::Internal("c", &[], r#"123"#),
+        ("a", [].as_slice(), "1"),
+        ("b", [].as_slice(), r#"b().wibble"#),
+        ("c", [].as_slice(), r#"123"#),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["a"], vec!["b"], vec!["c"]]
     );
 }
@@ -216,12 +197,12 @@ fn record_access() {
 #[test]
 fn binop() {
     let functions = [
-        Input::Internal("a", &[], r#"1 + a() + 2 / b() * 4"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"1 + a() + 2 / b() * 4"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["a"], vec!["c"]]
     );
 }
@@ -229,12 +210,12 @@ fn binop() {
 #[test]
 fn bit_strings() {
     let functions = [
-        Input::Internal("a", &[], r#"<<b, c>>"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"<<b, c>>"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["c"], vec!["a"]]
     );
 }
@@ -242,12 +223,12 @@ fn bit_strings() {
 #[test]
 fn tuple_index() {
     let functions = [
-        Input::Internal("a", &[], r#"b.0"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"b.0"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["a"], vec!["c"]]
     );
 }
@@ -255,12 +236,12 @@ fn tuple_index() {
 #[test]
 fn record_update() {
     let functions = [
-        Input::Internal("a", &[], r#"Wibble(..b, wobble: c())"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"Wibble(..b, wobble: c())"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["c"], vec!["a"]]
     );
 }
@@ -268,12 +249,12 @@ fn record_update() {
 #[test]
 fn negate() {
     let functions = [
-        Input::Internal("a", &[], r#"!c()"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"!c()"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["c"], vec!["a"]]
     );
 }
@@ -281,12 +262,12 @@ fn negate() {
 #[test]
 fn use_() {
     let functions = [
-        Input::Internal("a", &[], r#"use x <- c"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"use x <- c"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["c"], vec!["a"]]
     );
 }
@@ -294,25 +275,25 @@ fn use_() {
 #[test]
 fn use_shadowing() {
     let functions = [
-        Input::Internal("a", &[], r#"123"#),
-        Input::Internal("b", &[], r#"{ use c <- a c }"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"123"#),
+        ("b", [].as_slice(), r#"{ use c <- a c }"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["a"], vec!["b"], vec!["c"]]
     );
 }
 
 #[test]
 fn fn_argument_shadowing() {
-    let functions = [
-        Input::Internal("a", &[], r#"fn(b) { c b }"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+    let functions = &[
+        ("a", [].as_slice(), r#"fn(b) { c b }"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["c"], vec!["a"]]
     );
 }
@@ -320,12 +301,12 @@ fn fn_argument_shadowing() {
 #[test]
 fn fn_argument_shadowing_then_not() {
     let functions = [
-        Input::Internal("a", &[], r#"{ fn(b) { c b } b }"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"{ fn(b) { c b } b }"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["c"], vec!["a"]]
     );
 }
@@ -333,49 +314,61 @@ fn fn_argument_shadowing_then_not() {
 #[test]
 fn let_var() {
     let functions = [
-        Input::Internal("a", &[], r#"{ let c = b c }"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"{ let c = b c }"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["a"], vec!["c"]]
     );
 }
 
 #[test]
 fn pattern_int() {
-    let functions = [Input::Internal("a", &[], r#"{ let 1 = x }"#)];
-    assert_eq!(parse_and_order(&functions).unwrap(), vec![vec!["a"]]);
+    let functions = [("a", [].as_slice(), r#"{ let 1 = x }"#)];
+    assert_eq!(
+        parse_and_order(functions.as_slice()).unwrap(),
+        vec![vec!["a"]]
+    );
 }
 
 #[test]
 fn pattern_float() {
-    let functions = [Input::Internal("a", &[], r#"{ let 1.0 = x }"#)];
-    assert_eq!(parse_and_order(&functions).unwrap(), vec![vec!["a"]]);
+    let functions = [("a", [].as_slice(), r#"{ let 1.0 = x }"#)];
+    assert_eq!(
+        parse_and_order(functions.as_slice()).unwrap(),
+        vec![vec!["a"]]
+    );
 }
 
 #[test]
 fn pattern_string() {
-    let functions = [Input::Internal("a", &[], r#"{ let "1.0" = x }"#)];
-    assert_eq!(parse_and_order(&functions).unwrap(), vec![vec!["a"]]);
+    let functions = [("a", [].as_slice(), r#"{ let "1.0" = x }"#)];
+    assert_eq!(
+        parse_and_order(functions.as_slice()).unwrap(),
+        vec![vec!["a"]]
+    );
 }
 
 #[test]
 fn pattern_underscore() {
-    let functions = [Input::Internal("a", &[], r#"{ let _ = x }"#)];
-    assert_eq!(parse_and_order(&functions).unwrap(), vec![vec!["a"]]);
+    let functions = [("a", [].as_slice(), r#"{ let _ = x }"#)];
+    assert_eq!(
+        parse_and_order(functions.as_slice()).unwrap(),
+        vec![vec!["a"]]
+    );
 }
 
 #[test]
 fn pattern_concat() {
     let functions = [
-        Input::Internal("a", &[], r#"{ let "a" <> c = b c }"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"{ let "a" <> c = b c }"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["a"], vec!["c"]]
     );
 }
@@ -383,12 +376,12 @@ fn pattern_concat() {
 #[test]
 fn pattern_tuple() {
     let functions = [
-        Input::Internal("a", &[], r#"{ let #(a, c) = b a c }"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"{ let #(a, c) = b a c }"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["a"], vec!["c"]]
     );
 }
@@ -396,12 +389,12 @@ fn pattern_tuple() {
 #[test]
 fn pattern_list() {
     let functions = [
-        Input::Internal("a", &[], r#"{ let [a, c] = b a c }"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"{ let [a, c] = b a c }"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["a"], vec!["c"]]
     );
 }
@@ -409,12 +402,12 @@ fn pattern_list() {
 #[test]
 fn pattern_list_spread() {
     let functions = [
-        Input::Internal("a", &[], r#"{ let [a, ..c] = b a c }"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"{ let [a, ..c] = b a c }"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["a"], vec!["c"]]
     );
 }
@@ -422,12 +415,16 @@ fn pattern_list_spread() {
 #[test]
 fn pattern_bit_string_segment_size_var_usage() {
     let functions = [
-        Input::Internal("a", &[], r#"{ let <<y:size(b), _:unit(3)>> = c y }"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        (
+            "a",
+            [].as_slice(),
+            r#"{ let <<y:size(b), _:unit(3)>> = c y }"#,
+        ),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["c"], vec!["a"]]
     );
 }
@@ -435,12 +432,12 @@ fn pattern_bit_string_segment_size_var_usage() {
 #[test]
 fn pattern_assign() {
     let functions = [
-        Input::Internal("a", &[], r#"{ let 1 as b = c b }"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"{ let 1 as b = c b }"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["c"], vec!["a"]]
     );
 }
@@ -448,12 +445,12 @@ fn pattern_assign() {
 #[test]
 fn pattern_constructor() {
     let functions = [
-        Input::Internal("a", &[], r#"{ let Ok(b) = c b }"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"{ let Ok(b) = c b }"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["c"], vec!["a"]]
     );
 }
@@ -461,12 +458,12 @@ fn pattern_constructor() {
 #[test]
 fn scope_reset() {
     let functions = [
-        Input::Internal("a", &[], r#"{ let x = { let b = 1 b } b }"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"{ let x = { let b = 1 b } b }"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["a"], vec!["c"]]
     );
 }
@@ -474,12 +471,12 @@ fn scope_reset() {
 #[test]
 fn case_subject() {
     let functions = [
-        Input::Internal("a", &[], r#"case b { _ -> 1 }"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"case b { _ -> 1 }"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["a"], vec!["c"]]
     );
 }
@@ -487,12 +484,12 @@ fn case_subject() {
 #[test]
 fn case_subjects() {
     let functions = [
-        Input::Internal("a", &[], r#"case b, c { _, _ -> 1 }"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"case b, c { _, _ -> 1 }"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["c"], vec!["a"]]
     );
 }
@@ -500,12 +497,12 @@ fn case_subjects() {
 #[test]
 fn case_pattern_shadow() {
     let functions = [
-        Input::Internal("a", &[], r#"case 1 { b -> b }"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"case 1 { b -> b }"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["a"], vec!["b"], vec!["c"]]
     );
 }
@@ -513,12 +510,12 @@ fn case_pattern_shadow() {
 #[test]
 fn case_use_in_clause() {
     let functions = [
-        Input::Internal("a", &[], r#"case 1 { _ -> b }"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"case 1 { _ -> b }"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["a"], vec!["c"]]
     );
 }
@@ -526,12 +523,12 @@ fn case_use_in_clause() {
 #[test]
 fn case_clause_doesnt_shadow_later_clauses() {
     let functions = [
-        Input::Internal("a", &[], r#"case 1 { b -> 1 _ -> b }"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"case 1 { b -> 1 _ -> b }"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["a"], vec!["c"]]
     );
 }
@@ -539,12 +536,12 @@ fn case_clause_doesnt_shadow_later_clauses() {
 #[test]
 fn case_clause_doesnt_shadow_after() {
     let functions = [
-        Input::Internal("a", &[], r#"{ case 1 { b -> 1 } b }"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"{ case 1 { b -> 1 } b }"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["a"], vec!["c"]]
     );
 }
@@ -552,12 +549,12 @@ fn case_clause_doesnt_shadow_after() {
 #[test]
 fn guard() {
     let functions = [
-        Input::Internal("a", &[], r#"case 1 { _ if b -> 1 }"#),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("a", [].as_slice(), r#"case 1 { _ if b -> 1 }"#),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["a"], vec!["c"]]
     );
 }
@@ -565,44 +562,44 @@ fn guard() {
 #[test]
 fn big_guard() {
     let functions = [
-        Input::Internal(
+        (
             "a",
-            &[],
+            [].as_slice(),
             r#"case 1 { _ if 1 == 2 || x != #(Ok(b), 123) -> 1 }"#,
         ),
-        Input::Internal("b", &[], r#"123"#),
-        Input::External("c"),
+        ("b", [].as_slice(), r#"123"#),
+        ("c", [].as_slice(), "1"),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["b"], vec!["a"], vec!["c"]]
     );
 }
 
 #[test]
 fn duplicate_external_function_name() {
-    let functions = [Input::External("c"), Input::External("c")];
-    _ = parse_and_order(&functions).unwrap_err();
+    let functions = [("c", [].as_slice(), "1"), ("c", [].as_slice(), "1")];
+    _ = parse_and_order(functions.as_slice()).unwrap_err();
 }
 
 #[test]
 fn duplicate_function_name() {
     let functions = [
-        Input::Internal("b", &[], r#"123456"#),
-        Input::Internal("b", &[], r#"123456"#),
+        ("b", [].as_slice(), r#"123456"#),
+        ("b", [].as_slice(), r#"123456"#),
     ];
-    _ = parse_and_order(&functions).unwrap_err();
+    _ = parse_and_order(functions.as_slice()).unwrap_err();
 }
 
 #[test]
 fn more_complex_cycle() {
     let functions = [
-        Input::Internal("a1", &[], r#"{ a2 }"#),
-        Input::Internal("a2", &[], r#"{ a3 a1 }"#),
-        Input::Internal("a3", &[], r#"{ a1 }"#),
+        ("a1", [].as_slice(), r#"{ a2 }"#),
+        ("a2", [].as_slice(), r#"{ a3 a1 }"#),
+        ("a3", [].as_slice(), r#"{ a1 }"#),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["a2", "a1"], vec!["a3"]]
     );
 }
@@ -610,11 +607,11 @@ fn more_complex_cycle() {
 #[test]
 fn function_argument_shadowing() {
     let functions = [
-        Input::Internal("a", &["b"], r#"b"#),
-        Input::Internal("b", &[], r#"Nil"#),
+        ("a", ["b"].as_slice(), r#"b"#),
+        ("b", [].as_slice(), r#"Nil"#),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["a"], vec!["b"]]
     );
 }
@@ -623,11 +620,11 @@ fn function_argument_shadowing() {
 #[test]
 fn bug_2275() {
     let functions = [
-        Input::Internal("one", &[], r#"two one"#),
-        Input::Internal("two", &[], r#"two"#),
+        ("one", [].as_slice(), r#"two one"#),
+        ("two", [].as_slice(), r#"two"#),
     ];
     assert_eq!(
-        parse_and_order(&functions).unwrap(),
+        parse_and_order(functions.as_slice()).unwrap(),
         vec![vec!["two"], vec!["one"]]
     );
 }