+++ compiler-core/src/erlang/tests.rs
@@ -22,7 +22,6 @@ mod patterns;
 mod pipes;
 mod records;
 mod reserved;
-mod statement_if;
 mod strings;
 mod todo;
 mod use_;
@@ -565,7 +564,7 @@ pub type Box {
 // https://github.com/gleam-lang/gleam/issues/2156
 #[test]
 fn dynamic() {
-    assert_erl!("pub external type Dynamic")
+    assert_erl!("pub type Dynamic")
 }
 
 // https://github.com/gleam-lang/gleam/issues/2166