+++ compiler-core/src/format.rs
@@ -2,7 +2,7 @@
 mod tests;
 
 use crate::{
-    ast::{CustomType, ExternalFunction, Function, Import, ModuleConstant, TypeAlias, Use, *},
+    ast::{CustomType, Function, Import, ModuleConstant, TypeAlias, Use, *},
     build::Target,
     docvec,
     io::Utf8Writer,
@@ -131,8 +131,8 @@ impl<'comments> Formatter<'comments> {
     }
 
     fn targetted_definition<'a>(&mut self, definition: &'a TargettedDefinition) -> Document<'a> {
-        let target = definition.target();
-        let definition = definition.inner();
+        let target = definition.target;
+        let definition = &definition.definition;
         let start = definition.location().start;
         let comments = self.pop_comments(start);
         let document = self.documented_definition(definition);
@@ -149,7 +149,7 @@ impl<'comments> Formatter<'comments> {
         let mut previous_was_import = false;
 
         for definition in &module.definitions {
-            let is_import = definition.inner().is_import();
+            let is_import = definition.definition.is_import();
 
             if documents.is_empty() {
                 // We don't insert empty lines before the first definition
@@ -221,24 +221,6 @@ impl<'comments> Formatter<'comments> {
                 opaque,
                 ..
             }) => self.custom_type(*public, *opaque, name, parameters, constructors, location),
-
-            Definition::ExternalFunction(ExternalFunction {
-                public,
-                arguments: args,
-                name,
-                return_: retrn,
-                module,
-                fun,
-                ..
-            }) => self
-                .external_fn_signature(*public, name, args, retrn)
-                .append(" =")
-                .append(line())
-                .append("  \"")
-                .append(module.as_str())
-                .append("\" \"")
-                .append(fun.as_str())
-                .append("\""),
 
             Definition::Import(Import {
                 module,
@@ -546,22 +528,6 @@ impl<'comments> Formatter<'comments> {
             .append(line().append(body).nest(INDENT).group())
             .append(line())
             .append("}")
-    }
-
-    pub fn external_fn_signature<'a, A>(
-        &mut self,
-        public: bool,
-        name: &'a str,
-        args: &'a [ExternalFnArg<A>],
-        retrn: &'a TypeAst,
-    ) -> Document<'a> {
-        pub_(public)
-            .to_doc()
-            .append("external fn ")
-            .append(name)
-            .append(self.external_fn_args(args))
-            .append(" -> ".to_doc().append(self.type_ast(retrn)))
-            .group()
     }
 
     fn expr_fn<'a>(
@@ -1186,16 +1152,6 @@ impl<'comments> Formatter<'comments> {
         }))
     }
 
-    fn external_fn_arg<'a, A>(&mut self, arg: &'a ExternalFnArg<A>) -> Document<'a> {
-        let comments = self.pop_comments(arg.location.start);
-        let doc = label(&arg.label).append(self.type_ast(&arg.annotation));
-        commented(doc.group(), comments)
-    }
-
-    fn external_fn_args<'a, A>(&mut self, args: &'a [ExternalFnArg<A>]) -> Document<'a> {
-        wrap_args(args.iter().map(|e| self.external_fn_arg(e)))
-    }
-
     fn call_arg<'a>(&mut self, arg: &'a CallArg<UntypedExpr>) -> Document<'a> {
         match &arg.label {
             Some(s) => commented(
@@ -1621,13 +1577,6 @@ impl<'a> Documentable<'a> for &'a UnqualifiedImport {
     }
 }
 
-fn label(label: &Option<SmolStr>) -> Document<'_> {
-    match label {
-        Some(s) => s.to_doc().append(": "),
-        None => nil(),
-    }
-}
-
 impl<'a> Documentable<'a> for &'a BinOp {
     fn to_doc(self) -> Document<'a> {
         match self {
@@ -1718,7 +1667,7 @@ fn bit_string<'a>(
         break_(",", ", ")
     };
     break_("<<", "<<")
-        .append(join(segments.into_iter(), comma))
+        .append(join(segments, comma))
         .nest(INDENT)
         .append(break_(",", ""))
         .append(">>")