+++ compiler-core/src/format/tests.rs
@@ -78,71 +78,6 @@ pub type Four
 }
 
 #[test]
-fn external_fn() {
-    assert_format!(
-        r#"external fn main() -> Int =
-  "app" "main"
-"#
-    );
-
-    assert_format!(
-        r#"external fn main() -> program.Exit =
-  "app" "main"
-"#
-    );
-
-    assert_format!(
-        r#"external fn main(List(String)) -> Int =
-  "app" "main"
-"#
-    );
-
-    assert_format!(
-        r#"external fn main(argv: List(String)) -> Int =
-  "app" "main"
-"#
-    );
-
-    assert_format!(
-        r#"external fn main(
-  a_really_long_argument_label: List(String),
-  another_really_long_argument_label: whatever,
-) -> Int =
-  "app" "main"
-"#
-    );
-
-    assert_format!(
-        r#"external fn main(
-  a_really_long_argument_label: List(String),
-  another_really_long_argument_label: whatever,
-) -> Container(
-  WowThisTypeHasJustTheLongestName,
-  WowThisTypeHasJustTheLongestName,
-  WowThisTypeHasJustTheLongestName,
-) =
-  "app" "main"
-"#
-    );
-
-    assert_format!(
-        r#"external fn erl_filter(
-  fn(key, value) -> Bool,
-  Map(key, value),
-) -> Map(key, value) =
-  "maps" "filter"
-"#
-    );
-
-    assert_format!(
-        r#"///
-external fn x(a, b, c) -> #(a, b, c) =
-  "" ""
-"#
-    );
-}
-
-#[test]
 fn type_alias() {
     assert_format!(
         "type Option(a) =
@@ -3276,7 +3211,10 @@ fn main() {
 }
 "
     );
+}
 
+#[test]
+fn doc_comments_1_test() {
     assert_format!(
         "/// one
 ///two
@@ -3285,29 +3223,41 @@ fn main() {
 }
 "
     );
+}
 
+#[test]
+fn doc_comments_2_test() {
     assert_format!(
         r#"/// one
 ///two
-external fn whatever() -> Nil =
-  "" ""
+@external(javascript, "", "")
+fn whatever() -> Nil
 "#
     );
+}
 
+#[test]
+fn doc_comments_3_test() {
     assert_format!(
         r#"/// one
 ///two
 type Thingy
 "#
     );
+}
 
+#[test]
+fn doc_comments_4_test() {
     assert_format!(
         r#"/// one
 ///two
 type Thingy
 "#
     );
+}
 
+#[test]
+fn doc_comments_5_test() {
     assert_format!(
         r#"/// one
 ///two
@@ -3316,7 +3266,10 @@ type Whatever {
 }
 "#
     );
+}
 
+#[test]
+fn doc_comments_6_test() {
     assert_format!(
         r#"/// one
 ///two
@@ -3324,7 +3277,10 @@ type Whatever =
   Int
 "#
     );
+}
 
+#[test]
+fn doc_comments_7_test() {
     assert_format!(
         r#"import one
 
@@ -3394,8 +3350,8 @@ fn comments5() {
     assert_format!(
         r#"// one
 //two
-external fn whatever() -> Nil =
-  "" ""
+@external(javascript, "", "")
+fn whatever() -> Nil
 "#
     );
 }
@@ -3695,7 +3651,10 @@ fn commented_fn_arguments() {
 }
 "
     );
+}
 
+#[test]
+fn commented_fn_arguments1() {
     assert_format!(
         "fn main(
   // comment1
@@ -3707,15 +3666,18 @@ fn commented_fn_arguments() {
 }
 "
     );
+}
 
+#[test]
+fn commented_fn_arguments2() {
     assert_format!(
-        "pub external fn main(
+        "@external(erlang, \"\", \"\")
+pub fn main(
   // comment1
   argument1: Type,
   // comment2
   argument2: Type,
-) -> Int =
-  \"\" \"\"
+) -> Int
 "
     );
 }