+++ compiler-core/src/type_/tests/errors.rs
@@ -623,7 +623,10 @@ pub type Person {
 pub fn get_name(person: Person) { person.name }
 pub fn get_age(person: Person) { person.age }"
     );
+}
 
+#[test]
+fn accessor_multiple_variants_multiple_positions2() {
     // We cannot access fields on custom types with multiple variants where they are in different positions e.g. 1st and 3rd
     assert_module_error!(
         "
@@ -753,22 +756,28 @@ fn module_could_not_unify12() {
 
 #[test]
 fn module_arity_error() {
-    assert_module_error!("external fn go(List(a, b)) -> a = \"\" \"\"");
+    assert_module_error!("fn go(x: List(a, b)) -> Int { 1 }");
 }
 
 #[test]
 fn module_private_type_leak_1() {
     assert_module_error!(
-        r#"external type PrivateType
-pub external fn leak_type() -> PrivateType = "" """#
+        r#"type PrivateType
+
+@external(erlang, "a", "b")
+pub fn leak_type() -> PrivateType
+"#
     );
 }
 
 #[test]
 fn module_private_type_leak_2() {
     assert_module_error!(
-        r#"external type PrivateType
-external fn go() -> PrivateType = "" ""
+        r#"type PrivateType
+
+@external(erlang, "a", "b")
+fn go() -> PrivateType
+
 pub fn leak_type() { go() }"#
     );
 }
@@ -776,8 +785,9 @@ pub fn leak_type() { go() }"#
 #[test]
 fn module_private_type_leak_3() {
     assert_module_error!(
-        r#"external type PrivateType
-external fn go() -> PrivateType = "" ""
+        r#"type PrivateType
+@external(erlang, "a", "b")
+fn go() -> PrivateType
 pub fn leak_type() { [go()] }"#
     );
 }
@@ -785,15 +795,16 @@ pub fn leak_type() { [go()] }"#
 #[test]
 fn module_private_type_leak_4() {
     assert_module_error!(
-        r#"external type PrivateType
-pub external fn go(PrivateType) -> Int = "" """#
+        r#"type PrivateType
+@external(erlang, "a", "b")
+pub fn go(x: PrivateType) -> Int"#
     );
 }
 
 #[test]
 fn module_private_type_leak_5() {
     assert_module_error!(
-        r#"external type PrivateType
+        r#"type PrivateType
 pub type LeakType { Variant(PrivateType) }"#
     );
 }
@@ -955,16 +966,21 @@ fn dupe(x) { x }"
 #[test]
 fn duplicate_function_names_4() {
     assert_module_error!(
-        "fn dupe() { 1 }
-external fn dupe(x) -> x = \"\" \"\""
+        r#"fn dupe() { 1 }
+@external(erlang, "a", "b")
+fn dupe(x) -> x
+"#
     );
 }
 
 #[test]
 fn duplicate_function_names_5() {
     assert_module_error!(
-        "external fn dupe(x) -> x = \"\" \"\"
-fn dupe() { 1 }"
+        r#"
+@external(erlang, "a", "b")
+fn dupe(x) -> x
+fn dupe() { 1 }
+"#
     );
 }
 
@@ -1042,18 +1058,22 @@ fn foo() { 2 }"
 #[test]
 fn duplicate_extfn_extfn() {
     assert_module_error!(
-        "external fn foo() -> Float =
-  \"module1\" \"function1\"
-external fn foo() -> Float =
-  \"module2\" \"function2\""
+        r#"
+@external(erlang, "module1", "function1")
+fn foo() -> Float
+@external(erlang, "module2", "function2")
+fn foo() -> Float
+"#
     );
 }
 
 #[test]
 fn duplicate_extfn_fn() {
     assert_module_error!(
-        "external fn foo() -> Float =
-  \"module1\" \"function1\"
+        "
+@external(erlang, \"module1\", \"function1\")
+fn foo() -> Float
+
 fn foo() { 2 }"
     );
 }
@@ -1062,8 +1082,10 @@ fn foo() { 2 }"
 fn duplicate_fn_extfn() {
     assert_module_error!(
         "fn foo() { 1 }
-external fn foo() -> Float =
-  \"module2\" \"function2\""
+
+@external(erlang, \"module2\", \"function2\")
+fn foo() -> Float
+"
     );
 }
 
@@ -1071,16 +1093,20 @@ external fn foo() -> Float =
 fn duplicate_const_extfn() {
     assert_module_error!(
         "const foo = 1
-external fn foo() -> Float =
-  \"module2\" \"function2\""
+
+@external(erlang, \"module2\", \"function2\")
+fn foo() -> Float
+"
     );
 }
 
 #[test]
 fn duplicate_extfn_const() {
     assert_module_error!(
-        "external fn foo() -> Float =
-  \"module1\" \"function1\"
+        "
+@external(erlang, \"module1\", \"function1\")
+fn foo() -> Float
+
 const foo = 2"
     );
 }
@@ -1345,13 +1371,23 @@ fn type_holes1() {
 #[test]
 fn type_holes2() {
     // Type holes cannot be used when decaring types or external functions
-    assert_module_error!(r#"external fn main() -> List(_) = "" """#);
+    assert_module_error!(
+        r#"
+@external(erlang, "a", "b")
+fn main() -> List(_)
+"#
+    );
 }
 
 #[test]
 fn type_holes3() {
     // Type holes cannot be used when decaring types or external functions
-    assert_module_error!(r#"external fn main(List(_)) -> Nil = "" """#);
+    assert_module_error!(
+        r#"
+@external(erlang, "a", "b")
+fn main(x: List(_)) -> Nil
+"#
+    );
 }
 
 #[test]